requirements.txt
beautifulsoup4==4.12.2
black==23.3.0
blinker==1.6.2
cachelib==0.10.2
cachetools==5.3.1
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
config==0.5.1
cryptography==41.0.1
cssselect==1.2.0
docutils==0.20.1
elastic-transport==8.4.0
elasticsearch==7.12.0
feedfinder2==0.0.4
feedparser==6.0.10
filelock==3.12.0
Flask==2.3.2
Flask-JWT-Extended
Flask-Session==0.5.0
idna==3.4
itsdangerous==2.1.2
jieba3k==0.35.1
Jinja2==3.1.2
joblib==1.2.0
lockfile==0.12.2
lxml==4.9.2
MarkupSafe==2.1.3
mypy-extensions==1.0.0
mysql-connector-python==8.0.33
newspaper3k==0.2.8
nltk==3.8.1
packaging==23.1
pathspec==0.11.1
pid==3.0.4
Pillow==9.5.0
platformdirs==3.5.1
protobuf==3.20.3
pycparser==2.21
PyJWT==2.7.0
PyMySQL==1.0.3
python-daemon==3.0.1
python-dateutil==2.8.2
python-dotenv==1.0.0
PyYAML==6.0
redis==4.5.5
regex==2023.6.3
requests==2.31.0
requests-file==1.5.1
schedule==1.2.0
setproctitle==1.3.2
sgmllib3k==1.0.0
six==1.16.0
soupsieve==2.4.1
SQLAlchemy==2.0.15
tinysegmenter==0.3
tldextract==3.4.4
tqdm==4.65.0
typing_extensions==4.6.3
urllib3==1.26.16
Werkzeug==2.3.4
